Changed the structure of differential polynomials
We extended the ideas for differential polynomials to allow any ring with operator.
We have included factories for difference and differential polynomial rings.
We have unified the functionality of the polynomial classes: the specific operations must be inherited from the category, making only one class neccessary to represent the elements of polynomial rings for differential, difference and generic case.