Heavy improvement in the base code (#7)
* Added file for differential rings

Added the main file for the structure of differential rings and included the main comments to remember properly what to add in the file.

* Added classes a basic definitions

Added the class for the category DifferentialRings. This class includes the definition of the methods that has to be implemented.
Next step: make the ring of differential polynomials work as an element in DifferentialRings

* Fixed the __init__.py files to only import necessary things

Now the import of the whole module provides the objects they are suppose to know.

Other objects can be imported as usual.

Fixed the tests to simplify the examples

* added submodule differential_ring to the installation

* Removed unused imports and fixing some typos

* Changed definition of required methods for parent in differential rings

* Added _derivative of elements as abstract method

* Adapted the class DiffPolynomial to have appropriate definitions of derivative and _derivative

Remove now unused imports

* Added the category DifferentialRing to differential polynomials

The cast to differential polynomials is now done using the element_class class which inherits from the element_class of the category DifferentialRings

* ...

* Added method for getting orders

Added a method for getting the order for each variable in a
differential polynomial and also a method to implement the syntax "in" for differential variables

* Added DifferentialSystem

Added a class for differential systems. This is a list of differential polynomials with some extra functionalities.

Implemented the builder and the checker for conditions SP1 and SP2.

* Fixed tests for this file

* Fixed some typos in names and doc

* Added method to get the algebgraic system from the differential system

* Improved the method ``algebraic_equations``

Now the methdo takes into account the varables of the system and puts them in a higher priority in the hierarchy.

* Implemented a simple procedure to compute the differential resultant of a system

* Changed build_sp1

Now the method build_sp1 is also called "extend_by_derivation". All the appearances has been changed and the name "build_sp1" is simply an alias to refer to the property.

* Added the option to add a specific derivation

Added the chance to select the desired derivation of the base ring for ring of differential polynomials.
We changed the __init__ to allow this new behavior and also added the corresponding checking in the Factory.

* Fixed sp2 for homogeneous systems

The condition sp2 failed to compute a resultant when the sytem was homogeneous. Now, we added a method to check this new condition and "is_sp2" now returns True whenever the resultant can be computed directly from the system.

* Added notebboks

Added notebooks with different examples of usage for the
ring of differential polynomials

* Slight improvmenet defining the derivation
Now the derivation is an element (if possible) of the derivation module of the base ring.

* Created generic category for Ring with Operator

Created a category to represent a ring together with a operator (R,d)

Extended the category to differential rings.

Implemented a basic wrappring classes for the normal and differential case.

* Added the equivalent for the difference case

* Changing derivative to derivation in parent structures

* Adapted the file of differential_polynomial_ring
Included the new methods from the categories.

* Removed old import

* Adapted the ring of differential polynomials to the new category

Now it make the base ring always a DifferentialRing

* Fixed some representations to adapt better to the base differential ring

* Improved the coercion and equality system

* Added coercion system
Removed some redundant code

* Updated setup file

* Updated setup

* Fixed some tests

* Adapted the file for the new differential ring structure

Fixed the tests

* Fixing tests

Changed the factory: now it checks the base is a DifferentialRing in the factory instead of the __init__ method.
This prevents some duplication and non-unicity of the class.

* Fixed coercion system to run other examples

* Changed the structure of differential polynomials

We extended the ideas for differential polynomials to allow any ring with operator.
We have included factories for difference and differential polynomial rings.
We have unified the functionality of the polynomial classes: the specific operations must be inherited from the category, making only one class neccessary to represent the elements of polynomial rings for differential, difference and generic case.

* Fixed a casting error: included method _is_valid_homomorphism_

* Fixed setup and code to pass the tests
Adaptes the tests to the new repr methods

* Adapted the differential systems to the new framework

Now the systems can be of any pinfinite polynomial ring over a ring with an operation.

Special classes have been included for the specific cases of Differencial and Difference systems.

All algorithms currently working were moved to the general class

* Improved the implementation of random_element

* Added methods degree, initial and separant

* Improved the random_elmeent method

* Added method to_operator

Slight change in method degree: now allows no generator and returns the total degree.

Method to_operator relates the linear polynomials with ore operators (if possible)

* Improvement of _element_constructor_

Now we can reverse the method to_operator with the normal constructor of the ring of infinite polynomials.

Small update in the __call__ method.

* Fixed typo in _has_coerce_map_from for ring_w_operator

* Updated version and documentation files

* Fixed the structure for documentation

Fixed some errors in the category parts of the code

* Removed the method degree --> falling back to previous implementation

* Added more todos for future work

* Added ore_algebra to the install requirements

Fixed the pylint call withing sage

* Fixed all comments from pylint