Added methods to extract subsystems and equations
Changed the property equations to _equations (so it does not ollide with the next method.

Added method `equation`, which returns an equation of the system (maybe after applying the operation).

Added method `equations`, which returns a list of equations.

Added method `subsystem` which creates a new system using the equations given by indices and put as variables a new list.

Added method `change_variables` that creates a new system with the same equations but with different set of variables.

Change the use of old property equations to the new methods (either equation or equations)