Aded method is_lienar to a system
The method ``is_linear`` of a system checks whether a whole system is linear in the given variables of the system.

Change method ``is_lienar`` in RWOPolynomial. Now it allows to provide
a list of variables and will check whether the polynomial
is linear in those specific variables.
A default value of ``None`` is left to preserve compatibility with previous versions.