Milestone version 0.0.4 (#22)
Finishing all issues of milestone 0.0.4

* Saving small changes

* Started a new reconstruction of module ring_w_operator

* Removed some unused imports

Preparing the implementation of the factory

* Added version method to __init__

* Complete remake of RingWithOperators

* Reorganize ring_w_operator module

Now it is all in one file.

* Started revision of diff_polynomial module

* Reviewed file diff_polynomial_element.py

* Updated __init__s to allow implemented structures easier

* Added iterator for the index bijection

* Finished review of the file diff_polynomial.ring.py

* Changed names for diff_polynomial to rwo_polynomial

Now the implementation is much more general. Hence the previous name could be misleading.

* Fixed the factorys of rings with operators

* Fixed setup and version for this issue and branch

* Added more tests and fixed several bugs

* Added the skew computation everytime we could and remove mentions to commutation.

Now we do not require (here) that the operators commute. The commutation methods are probabilisitic and have high probabilities of returning False when it should. But it may be times this is not the case and the code will not detect it.

* Added methods for constants

* Fixed tests

Added command in Makefile to clean build files

* Change interface to apply operators

Now, at elements, the order of argumetns is reversed (first operation, then times).
Also, None is the default operation and only works when only one operator is available.

* Fixed tests, documentation

Added method "next" to generators to unify the code of getting next variable after applying an operator
Fixed method "contains"

* Fixed and added more tests

Checked everything (except evaluation) works.

Set only one factory and created methods for specific cases (but using the same factory for caching)

Fixed bug when caching the computation of operation in operator polynomials.

* Added evaluation of infinite polynomials

* Passed linting over the module

* Fixed some typos in grammar

* Improved repr and latex for rings

Passed linting over these files
Passed tests on these files

* Adapted system file

* Improved printing of identity homomorphisms

* Fixed method in differential resultant

* Added rwo_polynomial.rwo_polynomial_system to the default import

* Fixed import * system throughout the repository

* Checked changes in the notebooks. There are some issues to be fixed

* Updated structure of documentation files

* Updated configuration of documentation

* Updated main site for the documentation

Moved README into __init__
Changed from md to docstring format

* Updated some documentation

Changed name of "all_commute" to "all_operators_commute"

* Fixed typos in documentation

* Adapted setup file and requirements file for future usage

* Removed timeout contextmanager: use the library timeoutcontext

* Fixed names of polynomials to allow algebraic conversion
Fixed the "index" method

* Fixed cell with timeout

* Fixed documentation

* Removed file dixon.py: no use and no content for it

* Fixe linting comments and unsuccessful tests

* Upgraded to version 0.0.4